finish

+++ Part4/bloglist/controllers/blogs.js
@@ -1,6 +1,5 @@
 const blogRouter = require("express").Router();
 const Blog = require("../models/blog");
-const User = require("../models/user");
 
 require("express-async-errors");
 
@@ -15,15 +14,15 @@ blogRouter.get("/", async (request, response) => {
 
 blogRouter.post("/", async (request, response) => {
   const body = request.body;
+  const user = request.user;
 
   if (!body.title || !body.url) {
     return response.status(400).json({ error: "title or url missing" });
   }
 
-  const user = await User.findOne({});
   if (!user) {
     return response.status(400).json({ error: "no user found" });
-  } 
+  }
 
   const blog = new Blog({
     title: body.title,
@@ -41,7 +40,11 @@ blogRouter.post("/", async (request, response) => {
 });
 
 blogRouter.get("/:id", async (request, response) => {
-  const blog = await Blog.findById(request.params.id);
+  const blog = await Blog.findById(request.params.id).populate("user", {
+    username: 1,
+    name: 1,
+    id: 1,
+  });
   if (blog) {
     response.json(blog);
   } else {
@@ -50,6 +53,17 @@ blogRouter.get("/:id", async (request, response) => {
 });
 
 blogRouter.delete("/:id", async (request, response) => {
+  const user = request.user;
+
+  const blog = await Blog.findById(request.params.id);
+  if (!blog) {
+    return response.status(404).json({ error: "blog not found" });
+  }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(403).json({ error: "forbidden" });
+  }
+
   await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 });
